pyproject.toml
[tool.poetry]
name = "csv2cdx"
version = "2.0.0"
description = "CLI tool to convert excel and csv files to CycloneDX JSON files"
authors = ["Richard_E <restwick@cybeats.com>", "Cybeats <cs@cybeats.com"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/cybeats/csv2cdx"

keywords = ["SBOM", "CycloneDX", "Software Bill of Materials"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^2.0.2"
cyclonedx-python-lib = "^7.4.0"
packageurl-python = "^0.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
csv2cdx = "csv2cdx.__main__:main"

